+++ src/main/java/chatcat/chatcatexception/ChatCatException.java
@@ -1,0 +1,14 @@
+package chatcat.chatcatexception;
+
+/**
+ * Exception handling for ChatCat
+ */
+public class ChatCatException extends Exception {
+
+    /**
+     * Returns a {@code ChatCatException} with the given message.
+     */
+    public ChatCatException(String str) {
+        super(str);
+    }
+}
