+++ src/main/java/chatcat/tasklist/TaskList.java
@@ -1,26 +1,39 @@
-package duke.tasklist;
+package chatcat.tasklist;
 
 import java.util.ArrayList;
 
-import duke.tasks.Deadline;
-import duke.tasks.Event;
-import duke.util.UI;
-import duke.tasks.Task;
-import duke.tasks.Todo;
+import chatcat.tasks.Deadline;
+import chatcat.tasks.Event;
+import chatcat.util.UI;
+import chatcat.tasks.Task;
+import chatcat.tasks.Todo;
+import chatcat.util.WriteToFile;
+import chatcat.chatcatexception.ChatCatException;
 
-import duke.util.WriteToFile;
-import duke.dukeexception.DukeException;
-
+/**
+ * Handles writing and reading to list of tasks represented as an {@code ArrayList}.
+ * Allows clients to unmark, mark, add and delete tasks. Allows users to view task list.
+ */
 public class TaskList {
     ArrayList<Task> Tasks = new ArrayList<>();
     WriteToFile writeToFile;
     UI ui;
 
+    /**
+     * Creates {@code TaskList} object containing an empty Task List {@code ArrayList}.
+     * Creates {@code UI} object.
+     */
     public TaskList() {
         ui = new UI();
         writeToFile = new WriteToFile();
     }
 
+    /**
+     * Prints out all tasks in Task List.
+     *
+     * @see WriteToFile
+     * @see Task
+     */
     public void listTasks() {
         Tasks = writeToFile.toRead();
 
@@ -35,6 +48,14 @@ public class TaskList {
         System.out.println("");
     }
 
+    /**
+     * Marks task at specified location.
+     *
+     * @param str containing the task to be marked.
+     * @see UI
+     * @see Task
+     * @see WriteToFile
+     */
     public void mark(String str) {
         String[] input = str.split(" ");
         int taskID = Integer.parseInt(input[1]) - 1;
@@ -44,6 +65,14 @@ public class TaskList {
                 Tasks.get(taskID));
     }
 
+    /**
+     * Unmark task at specified location.
+     *
+     * @param str containing the task to be marked.
+     * @see UI
+     * @see Task
+     * @see WriteToFile
+     */
     public void unmark(String str) {
         String[] input = str.split(" ");
         int taskID = Integer.parseInt(input[1]) - 1;
@@ -53,10 +82,19 @@ public class TaskList {
                 Tasks.get(taskID));
     }
 
-    public void setTodo(String str) throws DukeException {
+    /**
+     * Creates a todo {@code todo} object and appends the object at the end of {@code taskList}.
+     *
+     * @param str containing the task to be added.
+     * @throws ChatCatException if description of todo is empty.
+     * @see UI
+     * @see Todo
+     * @see WriteToFile
+     */
+    public void setTodo(String str) throws ChatCatException {
         String[] input = str.split(" ");
         if (input.length == 1) {
-            throw new DukeException("OOPS!!! The description of a todo cannot be empty.");
+            throw new ChatCatException("OOPS!!! The description of a todo cannot be empty.");
         }
 
         Todo todo = new Todo(str.substring(5));
@@ -66,11 +104,20 @@ public class TaskList {
         ui.printOutPut("Now you have " + Tasks.size() + " tasks in the list.");
     }
 
-    public void setDeadline(String str) throws DukeException {
+    /**
+     * Creates a deadline {@code Deadline} object and appends the object at the end of {@code taskList}.
+     *
+     * @param str containing the task to be added.
+     * @throws ChatCatException if description of deadline is empty.
+     * @see UI
+     * @see Deadline
+     * @see WriteToFile
+     */
+    public void setDeadline(String str) throws ChatCatException {
         String[] input = str.split(" ");
 
         if (input.length == 1) {
-            throw new DukeException("OOPS!!! The description of a deadline cannot be empty.");
+            throw new ChatCatException("OOPS!!! The description of a deadline cannot be empty.");
         }
 
         String[] split = str.split("/by ");
@@ -81,11 +128,20 @@ public class TaskList {
         ui.printOutPut("Now you have " + Tasks.size() + " tasks in the list.");
     }
 
-    public void setEvent(String str) throws DukeException {
+    /**
+     * Creates a deadline {@code Event} object and appends the object at the end of {@code taskList}.
+     *
+     * @param str containing the task to be added.
+     * @throws ChatCatException if description of event is empty.
+     * @see UI
+     * @see Event
+     * @see WriteToFile
+     */
+    public void setEvent(String str) throws ChatCatException {
         String[] input = str.split(" ");
 
         if (input.length == 1) {
-            throw new DukeException("OOPS!!! The description of a event cannot be empty.");
+            throw new ChatCatException("OOPS!!! The description of a event cannot be empty.");
         }
 
         String[] split = str.split("/at ");
@@ -96,6 +152,14 @@ public class TaskList {
         ui.printOutPut("Now you have " + Tasks.size() + " tasks in the list." + "\n");
     }
 
+    /**
+     * Deletes a specified task {@code Task} in the tasklist {@code taskList}.
+     *
+     * @param str containing the task to be deleted.
+     * @see UI
+     * @see Task
+     * @see WriteToFile
+     */
     public void delete(String str) {
         Tasks = writeToFile.toRead();
         String[] input = str.split(" ");
