+++ src/main/java/chatcat/tasks/Event.java
@@ -1,10 +1,17 @@
-package duke.tasks;
+package chatcat.tasks;
 
 import java.time.DateTimeException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 
+/**
+ * The default Event class. Inherited from {@code Task}.
+ *
+ * @see Task
+ * @see DateTimeFormatter
+ * @see DateTimeFormatterBuilder
+ */
 public class Event extends Task {
     String event;
     LocalDate time;
@@ -20,7 +27,14 @@ public class Event extends Task {
 
     DateTimeFormatter dateTimeFormatter = dateTimeFormatterBuilder.toFormatter();
 
-
+    /**
+     * Creates a {@code Event} object using a specified description.
+     *
+     * @param event the description of this task.
+     * @param time the time of the task.
+     * @throws DateTimeException if user inputs an invalid time
+     * @see LocalDate
+     */
     public Event(String event, String time) throws DateTimeException {
         super(event);
         this.event = event;
@@ -33,6 +47,11 @@ public class Event extends Task {
         }
     }
 
+    /**
+     * Returns a representation in string of {@code Event} task.
+     *
+     * @return a representation in string of {@code Event} task.
+     */
     @Override
     public String toString() {
         String temp = time.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
