--- src/main/java/duke/dukeexception/DukeException.java
@@ -1,5 +1,0 @@
-package duke.dukeexception;
-
-public class DukeException extends Exception {
-    public DukeException(String str) { super(str); }
-}
