--- src/main/java/duke/tasks/Todo.java
@@ -1,12 +1,0 @@
-package duke.tasks;
-
-public class Todo extends Task {
-    public Todo(String todo) {
-        super(todo);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
