+++ src/test/java/ChatCatTest.java
@@ -2,7 +2,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class DukeTest {
+public class ChatCatTest {
     @Test
     public void dummyTest(){
         assertEquals(2, 2);
